training/requirements.txt
torch>=2.1.0
miditok>=3.0.0
symusic>=0.5.0
huggingface-hub>=0.20.0
wandb>=0.16.0
tqdm>=4.65.0
pyyaml>=6.0
numpy>=1.24.0
onnx>=1.15.0
onnxruntime>=1.16.0
